pyproject.toml
[project]
name = "agentic_kg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andreas Kollegger", email = "akollegger@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "clevercsv[full]>=0.8.3",
    "google-adk>=1.10.0",
    "litellm>=1.75.5.post1",
    "neo4j>=5.28.2",
    "neo4j-graphrag>=1.9.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
]

[project.scripts]
agentic_kg = "agentic_kg:main"

[build-system]
requires = ["uv_build>=0.8.7,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.2.0",
    "testcontainers>=4.7.2",
]

# Pytest defaults for convenience
[tool.pytest.ini_options]
addopts = "-q -m 'not integration'"
testpaths = ["tests"]
markers = [
    "integration: mark tests that require external services such as Docker/Neo4j",
]
